refactor: move constant
