+++ packages/universal-router-sdk/src/index.ts
@@ -1,4 +1,4 @@
-export { PancakeSwapUniversalRouter } from './swapRouter'
+export { ROUTER_AS_RECIPIENT, getUniversalRouterAddress } from './constants'
 export * from './entities'
 export * from './entities/types'
-export { getUniversalRouterAddress, ROUTER_AS_RECIPIENT } from './utils/constants'
+export { PancakeSwapUniversalRouter } from './swapRouter'
